+++ src/Hazelcast.Net.Tests/Clustering/OrphanSubscriptionTests.cs
@@ -24,7 +24,7 @@ using Hazelcast.Messaging;
 using Hazelcast.Networking;
 using Hazelcast.Protocol;
 using Hazelcast.Protocol.Codecs;
-using Hazelcast.Protocol.Data;
+using Hazelcast.Protocol.Models;
 using Hazelcast.Serialization;
 using Hazelcast.Testing;
 using Hazelcast.Testing.Protocol;
