+++ src/Hazelcast.Net.Tests/Exceptions/RemoteExceptionTests.cs
@@ -17,7 +17,7 @@ using System.Collections.Generic;
 using Hazelcast.Clustering;
 using Hazelcast.Exceptions;
 using Hazelcast.Protocol;
-using Hazelcast.Protocol.Data;
+using Hazelcast.Protocol.Models;
 using Hazelcast.Testing;
 using Hazelcast.Tests.Testing;
 using NUnit.Framework;
