+++ src/Hazelcast.Net/Protocol/CustomCodecs/AnchorDataListHolderCodec.cs
@@ -39,7 +39,7 @@ namespace Hazelcast.Protocol.CustomCodecs
     internal static class AnchorDataListHolderCodec
     {
 
-        public static void Encode(ClientMessage clientMessage, Hazelcast.Protocol.Data.AnchorDataListHolder anchorDataListHolder)
+        public static void Encode(ClientMessage clientMessage, Hazelcast.Protocol.Models.AnchorDataListHolder anchorDataListHolder)
         {
             clientMessage.Append(Frame.CreateBeginStruct());
 
@@ -49,7 +49,7 @@ namespace Hazelcast.Protocol.CustomCodecs
             clientMessage.Append(Frame.CreateEndStruct());
         }
 
-        public static Hazelcast.Protocol.Data.AnchorDataListHolder Decode(IEnumerator<Frame> iterator)
+        public static Hazelcast.Protocol.Models.AnchorDataListHolder Decode(IEnumerator<Frame> iterator)
         {
             // begin frame
             iterator.Take();
@@ -57,7 +57,7 @@ namespace Hazelcast.Protocol.CustomCodecs
             var anchorDataList = EntryListCodec.Decode(iterator, DataCodec.Decode, DataCodec.Decode);
 
             iterator.SkipToStructEnd();
-            return new Hazelcast.Protocol.Data.AnchorDataListHolder(anchorPageList, anchorDataList);
+            return new Hazelcast.Protocol.Models.AnchorDataListHolder(anchorPageList, anchorDataList);
         }
     }
 }
