+++ src/Hazelcast.Net/Protocol/ErrorsCodec.cs
@@ -16,7 +16,7 @@ using System.Collections.Generic;
 using Hazelcast.Messaging;
 using Hazelcast.Protocol.BuiltInCodecs;
 using Hazelcast.Protocol.CustomCodecs;
-using Hazelcast.Protocol.Data;
+using Hazelcast.Protocol.Models;
 
 namespace Hazelcast.Protocol
 {
