+++ src/Hazelcast.Net/Protocol/Models/PagingPredicateHolder.cs
@@ -17,7 +17,7 @@ using System.Collections.Generic;
 using Hazelcast.Query;
 using Hazelcast.Serialization;
 
-namespace Hazelcast.Protocol.Data
+namespace Hazelcast.Protocol.Models
 {
     internal class PagingPredicateHolder
     {
